ammend

+++ app/core/models/base_model.py
@@ -4,7 +4,8 @@ from datetime import datetime, date
 from decimal import Decimal
 from typing import Annotated, Optional, Type, List
 # from sqlalchemy.dialects.postgresql import MONEY
-from sqlalchemy import DateTime, DECIMAL, func, text, Text, Computed, inspect, BigInteger, ARRAY
+from sqlalchemy import DateTime, DECIMAL, func, text, Text, Computed, inspect
+from sqlalchemy.dialects.postgresql import ARRAY, BIGINT
 from sqlalchemy.dialects.postgresql import TSVECTOR
 # from sqlalchemy_serializer import SerializerMixin
 from sqlalchemy.ext.asyncio import AsyncAttrs
@@ -240,7 +241,7 @@ class Search:
     """ поисковое поле для триграммного индекса """
     __abstract__ = True
 
-    word_hashes: Mapped[List[int]] = mapped_column(ARRAY(BigInteger), nullable=True)
+    word_hashes: Mapped[List[int]] = mapped_column(ARRAY(BIGINT), nullable=True)
 
     @declared_attr
     def search_content(cls) -> Mapped[Optional[str]]:
